LibI4/Python_AI/requirements.txt
# General requirements
numpy
torch>=2.0.0
torchtext
torchvision
torchaudio
transformers~=4.42.3, >=4.42.3
diffusers~=0.29.0, >=0.29.0

# Server requirements
asyncio
websockets
pymysql~=1.1.0

# Other requirements
requests
beautifulsoup4
google
soundfile
scipy
opencv-python
Pillow
setuptools
safetensors
sentencepiece
emoji
google-tts
git+https://github.com/Tps-F/fairseq.git@main
bitsandbytes
#huggingface-hub>=0.19.0,<0.20.0
psutil

# Requirements for inference
gpt4all~=2.7.0
llama-cpp-python
SpeechRecognition~=3.10.4
openai-whisper>=20231117
git+https://github.com/RVC-Project/Retrieval-based-Voice-Conversion.git@develop
pySpeakNG~=0.1.2
#styletts2~=0.1.6

# Requirements for the Discord bot
discord.py